Rename class name: KafkaLZ4BlockOutputStream → Lz4OutputStream
